+++ crates/yuck/src/format_diagnostic.rs
@@ -22,6 +22,7 @@ pub fn span_to_secondary_label(span: Span) -> Label<usize> {
 ///     kind = Severity::Error,
 ///     msg = format!("Expected value, but got `{}`", actual),
 ///     label = span => "Expected some value here",
+///     note = "help: you can have many notes in the diagnostic",
 ///     note = format!("Got: {}", actual),
 /// }
 /// ```
@@ -30,7 +31,7 @@ macro_rules! gen_diagnostic {
     ( $(kind = $kind:expr,)?
       $(msg = $msg:expr)?
       $(, label = $span:expr $(=> $label:expr)?)?
-      $(, note = $note:expr)? $(,)?
+      $(, note = $note:expr)* $(,)?
     ) => {
         ::codespan_reporting::diagnostic::Diagnostic::new(gen_diagnostic! {
             @macro_fallback $({$kind})? {::codespan_reporting::diagnostic::Severity::Error}
@@ -40,7 +41,7 @@ macro_rules! gen_diagnostic {
                 ::codespan_reporting::diagnostic::Label::primary($span.2, $span.0..$span.1)
                     $(.with_message($label))?
             ]))?
-            $(.with_notes(vec![$note.to_string()]))?
+            .with_notes(vec![$($note.to_string(),)*])
     };
     ($msg:expr $(, $span:expr $(,)?)?) => {{
         ::codespan_reporting::diagnostic::Diagnostic::error()